requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
--index-url https://pypi.tuna.tsinghua.edu.cn/simple

aiofiles==23.2.1
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
autocut-sub @ git+https://github.com/methanal/autocut.git@libs
    # via -r requirements.in
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
decorator==4.4.2
    # via moviepy
distro==1.9.0
    # via openai
dnspython==2.6.1
    # via email-validator
email-validator==2.1.1
    # via fastapi
exceptiongroup==1.2.1
    # via anyio
fastapi==0.111.0
    # via
    #   -r requirements.in
    #   sentry-sdk
fastapi-cli==0.0.4
    # via fastapi
ffmpeg-python==0.2.0
    # via autocut-sub
filelock==3.15.1
    # via torch
fsspec==2024.6.0
    # via torch
future==1.0.0
    # via ffmpeg-python
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via
    #   -r requirements.in
    #   fastapi
    #   openai
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
imageio==2.34.1
    # via moviepy
imageio-ffmpeg==0.5.1
    # via moviepy
jinja2==3.1.4
    # via
    #   -r requirements.in
    #   fastapi
    #   torch
llvmlite==0.43.0
    # via numba
lz4==4.3.3
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.3.0
    # via openai-whisper
moviepy==1.0.3
    # via
    #   -r requirements.in
    #   autocut-sub
mpmath==1.3.0
    # via sympy
networkx==3.2.1
    # via torch
numba==0.60.0
    # via openai-whisper
numpy==1.26.4
    # via
    #   imageio
    #   moviepy
    #   numba
    #   openai-whisper
openai==1.34.0
    # via -r requirements.in
openai-whisper==20231117
    # via autocut-sub
opencc-python-reimplemented==0.1.7
    # via autocut-sub
orjson==3.10.5
    # via
    #   -r requirements.in
    #   fastapi
parameterized==0.9.0
    # via autocut-sub
pillow==10.3.0
    # via imageio
proglog==0.1.10
    # via moviepy
pydantic==2.7.4
    # via
    #   fastapi
    #   openai
    #   pydantic-settings
pydantic-core==2.18.4
    # via pydantic
pydantic-settings==2.3.3
    # via -r requirements.in
pydub==0.25.1
    # via autocut-sub
pygments==2.18.0
    # via rich
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.1
    # via uvicorn
regex==2024.5.15
    # via tiktoken
requests==2.32.3
    # via
    #   moviepy
    #   tiktoken
rich==13.7.1
    # via typer
sentry-sdk[fastapi]==2.5.1
    # via -r requirements.in
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
srt==3.5.3
    # via autocut-sub
starlette==0.37.2
    # via fastapi
sympy==1.12.1
    # via torch
tiktoken==0.7.0
    # via openai-whisper
torch==2.3.1
    # via
    #   openai-whisper
    #   torchaudio
torchaudio==2.3.1
    # via autocut-sub
tqdm==4.66.4
    # via
    #   autocut-sub
    #   moviepy
    #   openai
    #   openai-whisper
    #   proglog
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   starlette
    #   torch
    #   typer
    #   uvicorn
ujson==5.10.0
    # via fastapi
urllib3==2.2.1
    # via
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.30.1
    # via
    #   -r requirements.in
    #   fastapi
uvloop==0.19.0
    # via uvicorn
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
